Redundent files removed and code for ml algorithms improved
