+++ ravop/ml/utils.py
@@ -1,0 +1,8 @@
+from common import db
+from ravop.core import Op
+
+
+def get_ops_by_name(op_name, graph_id=None):
+    ops = db.get_ops_by_name(op_name=op_name, graph_id=graph_id)
+
+    return [Op(id=op.id) for op in ops]
